r12029/sofa-dev : some warnings cleaned
 //partial sync (8/9 changes)//


Former-commit-id: f024dbc31e8d32cd523b1cd987850c2e0040e9c0

+++ modules/sofa/component/collision/PointModel.cpp
@@ -254,7 +254,7 @@ bool PointModel::canCollideWithElement(int index, CollisionModel* model2, int in
 
                 if (debug)
                     std::cout<<"v1 = "<<v1 <<"  verticesAroundVertex2[i2]"<<verticesAroundVertex2[i2]<<std::endl;
-                if (v1==verticesAroundVertex2[i2] || v1==index2 || index == verticesAroundVertex2[i2])
+                if (v1==verticesAroundVertex2[i2] || v1==(unsigned int)index2 || index == (int)verticesAroundVertex2[i2])
                 {
                     if(debug)
                         std::cout<<" return false"<<std::endl;
